+++ app/src/main/java/com/leeseungyun1020/learningtrip/model/PlaceReview.kt
@@ -1,13 +1,13 @@
 package com.leeseungyun1020.learningtrip.model
 
 data class PlaceReview(
-    val placeId: Int,
+    val placeId: Long,
     val rating: Double,
     val review: Int,
 )
 
 data class SimplePlaceReview(
-    val placeId: Int,
+    val placeId: Long,
     val placeImageURL: String,
     val placeName: String,
     val rating: Double,
